news_monitoring_cloud_function/requirements.txt
google-cloud-firestore>=2.13.0
google-cloud-functions>=1.16.0
requests>=2.31.0
functions-framework>=3.5.0
